Initial commit
This text based adventure game was created as part of a final project for C++. It uses  a dice rolling mechanic by taking a number from user input to create stats. After stats are created, text based prompts come up and stats are rolled using a number generator from 1, stat. If the roll for the stat is lower than the number it passes. This was a lot of fun to make, I hope this helps anyone in C++.